Adjusted a few parts of the code that were causing errors
In the login-view component, I had the data object with the keys 'access' and 'secret' from the exercise, instead of 'username' and 'password' like it should have been. I adjusted the way the movie component is set back to how it was with the previous exercise (Authorization should still be there) since the way I have the data for my director and genre set up makes it so that I can't just do .then((movies) => { setMovies(movies); } since genre and director in my database are objects that hold an ID. I'll likely have to go back and change that at some point in the future to avoid similar problems - since they only hold one thing, there's no reason I can't just simplify it to a string instead of an object.